+++ NEWS.md
@@ -1,5 +1,20 @@
 # News
 
+## 2022-01
+
+### Changes in rocker_scripts
+
+- `userconf.sh`, which is executed when RStudio Server is launched, has been split into `init_set_env.sh` and `init_userconf.sh`. `init_set_env.sh` is also executed when Shiny Server is launched. ([#320](https://github.com/rocker-org/rocker-versioned2/pull/320))
+
+### Changes in pre-built images
+
+- Environment variables set in the container are now reflected in the sessions on Shiny Server. ([#320](https://github.com/rocker-org/rocker-versioned2/pull/320))
+- Change the date of the CRAN mirrors that were set for the R 4.0.0, R 4.0.1, R 4.0.2, and R 4.0.3 images. This reconfiguration was done mechanically, selecting the mirrors just before the release date of the next version of R. ([#328](https://github.com/rocker-org/rocker-versioned2/pull/328))
+  - R 4.0.0: 2020-06-08 to 2020-06-04 (The next version, R 4.0.1 was released on 2020-06-06)
+  - R 4.0.1: 2020-06-25 to 2020-06-18 (The next version, R 4.0.2 was released on 2020-06-22)
+  - R 4.0.2: 2020-10-07 to 2020-10-09 (The next version, R 4.0.3 was released on 2020-10-10)
+  - R 4.0.3: 2021-02-17 to 2021-02-11 (The next version, R 4.0.4 was released on 2021-02-15)
+
 ## 2021-12
 
 ### Changes in pre-built images
@@ -34,3 +49,10 @@
 
 - New builds of `rocker/verse` and `rocker/ml-verse` for R >= 4.1.0 will install [quarto-cli](https://github.com/quarto-dev/quarto-cli) and the quarto R package. ([#253](https://github.com/rocker-org/rocker-versioned2/pull/253))
 - The `install2.r` command has been replaced by rocker scripts version of `install2.r`. ([#275](https://github.com/rocker-org/rocker-versioned2/pull/275))
+
+## 2021-06
+
+### Changes in pre-built images
+
+- The time zone setting for R, which was hardcoded as `TZ=Etc/UTC` in `Renviron.site`, has been removed. The environment variable `TZ` set in the container will be used instead. ([#151](https://github.com/rocker-org/rocker-versioned2/issues/151), [#186](https://github.com/rocker-org/rocker-versioned2/pull/186))
+- Environment variables set in the container are now reflected in the sessions on RStudio Server. ([#186](https://github.com/rocker-org/rocker-versioned2/pull/186), [#189](https://github.com/rocker-org/rocker-versioned2/pull/189), [#190](https://github.com/rocker-org/rocker-versioned2/pull/190))